+++ src/components/App.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
-import { Prompt, BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Dashboard from '@pages/Dashboard';
 import Login from '@pages/Login';
+import Alert from '@components/Alert';
 import Header from '@components/Header';
 import Page from '@components/Page';
 import PrivateRoute from '@components/PrivateRoute';
@@ -11,29 +12,29 @@ import { GlobalStyle } from '@styles/global';
 
 const App: React.FC = (): JSX.Element => {
   const [isAuthorized, setIsAuthorized] = useState(false);
-  const [isLogoutEnabled, setIsLogoutEnabled] = useState(false);
+  const [shouldDisplayLogoutAlert, setShouldDisplayLogoutAlert] = useState(false);
 
   return (
     <ThemeProvider theme={Theme}>
       <GlobalStyle />
       <div className='app'>
         <Router>
-          <Header isAuthorized={isAuthorized} />
+          <Header
+            isAuthorized={isAuthorized}
+            setShouldDisplayLogoutAlert={setShouldDisplayLogoutAlert}
+          />
           <Page>
-            <Prompt message={(location): string | true =>
-              isLogoutEnabled && location.pathname === '/login'
-                ? 'Are you sure you want to log out?' : true}
-            />
+            {shouldDisplayLogoutAlert &&
+              <Alert
+                setIsAuthorized={setIsAuthorized}
+                setShouldDisplayLogoutAlert={setShouldDisplayLogoutAlert}
+              />
+            }
             <Switch>
               <Route
                 path="/login"
                 render={(): JSX.Element =>
-                  <Login
-                    isAuthorized={isAuthorized}
-                    isLogoutEnabled={isLogoutEnabled}
-                    setIsAuthorized={setIsAuthorized}
-                    setIsLogoutEnabled={setIsLogoutEnabled}
-                  />
+                  <Login setIsAuthorized={setIsAuthorized} />
                 }
               />
               <PrivateRoute isAuthorized={isAuthorized} path="*" component={Dashboard} />