+++ src/components/Header.tsx
@@ -4,15 +4,27 @@ import StyledHeader from '@styles/components/Header.styled';
 
 type HeaderProps = {
   readonly isAuthorized: boolean;
+  readonly setShouldDisplayLogoutAlert: (arg: boolean) => void;
 };
 
-const Header: React.FC<HeaderProps> = ({ isAuthorized }: HeaderProps): JSX.Element =>
-  <StyledHeader>
-    <h6><Link to='/dashboard'>TIMEWARD</Link></h6>
-    <nav>
-      {isAuthorized && <Link to='/dashboard'>Dashboard</Link>}
-      <Link to='/login'>{isAuthorized ? 'Logout' : 'Login'}</Link>
-    </nav>
-  </StyledHeader>;
+const Header: React.FC<HeaderProps> =
+  ({ isAuthorized, setShouldDisplayLogoutAlert }: HeaderProps): JSX.Element => {
+    const handleLogout = (): void => setShouldDisplayLogoutAlert(true);
+
+    return (
+      <StyledHeader>
+        <h6><Link to='/dashboard'>TIMEWARD</Link></h6>
+        <nav>
+          {isAuthorized &&
+            <>
+              <Link to='/dashboard'>Dashboard</Link>
+              <a href='#logout' onClick={handleLogout}>Logout</a>
+            </>
+          }
+          {!isAuthorized && <Link to='/login'>Login</Link>}
+        </nav>
+      </StyledHeader>
+    );
+  };
 
 export default Header;